Restricts build description to be no more than 100 characters
Elastic Beanstalk has a limitation for build description to be no more than
100 characters. If a build is submitted with more than then limitation, the
build request fails.

Since we are pulling the build description from the commit message, this limits
the commit message to be always within 100 characters. This makes restricts
the amount of useful information that we can place into the commit message.